refactor: move app/router generation from `main.rs` into a separate function in `lib.rs`
